[refactor] lostAnimal에서 lostPost로 디렉토리 및 클래스 이름 변경
