--- src/main/java/kr/co/pawong/pwbe/lostAnimal/adapter/in/api/LostAnimalQueryController.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.adapter.in.api;
-
-public class LostAnimalQueryController {
-
-}
