--- src/main/java/kr/co/pawong/pwbe/lostAnimal/application/port/in/LostAnimalQueryUseCase.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.application.port.in;
-
-public interface LostAnimalQueryUseCase {
-
-}
