--- src/main/java/kr/co/pawong/pwbe/lostAnimal/application/port/in/LostPostUpdateUseCase.java
@@ -1,7 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.application.port.in;
-
-import kr.co.pawong.pwbe.lostAnimal.domain.LostPost;
-
-public interface LostPostUpdateUseCase {
-    LostPost createLostPost(LostPost lostPost, Long userId);
-}
