--- src/main/java/kr/co/pawong/pwbe/lostAnimal/application/service/LostAnimalQueryService.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.application.service;
-
-public class LostAnimalQueryService {
-
-}
