--- src/main/java/kr/co/pawong/pwbe/lostAnimal/domain/LostAnimalUpdate.java
@@ -1,5 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.domain;
-
-public class LostAnimalUpdate {
-
-}
