--- src/main/java/kr/co/pawong/pwbe/lostAnimal/enums/PostStatus.java
@@ -1,6 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.enums;
-
-public enum PostStatus {
-    ACTIVE,
-    DELETED
-}
