--- src/main/java/kr/co/pawong/pwbe/lostAnimal/enums/PostType.java
@@ -1,7 +1,0 @@
-package kr.co.pawong.pwbe.lostAnimal.enums;
-
-public enum PostType {
-    LOST,
-    FOUND,
-    FOSTER
-}
