+++ src/main/java/kr/co/pawong/pwbe/lostPost/adapter/in/api/LostPostQueryController.java
@@ -1,0 +1,5 @@
+package kr.co.pawong.pwbe.lostPost.adapter.in.api;
+
+public class LostPostQueryController {
+
+}
