+++ src/main/java/kr/co/pawong/pwbe/lostPost/application/port/in/LostPostQueryUseCase.java
@@ -1,0 +1,5 @@
+package kr.co.pawong.pwbe.lostPost.application.port.in;
+
+public interface LostPostQueryUseCase {
+
+}
