+++ src/main/java/kr/co/pawong/pwbe/lostPost/application/port/in/LostPostUpdateUseCase.java
@@ -1,0 +1,7 @@
+package kr.co.pawong.pwbe.lostPost.application.port.in;
+
+import kr.co.pawong.pwbe.lostPost.domain.LostPost;
+
+public interface LostPostUpdateUseCase {
+    LostPost createLostPost(LostPost lostPost, Long userId);
+}
