+++ src/main/java/kr/co/pawong/pwbe/lostPost/application/service/LostPostQueryService.java
@@ -1,0 +1,5 @@
+package kr.co.pawong.pwbe.lostPost.application.service;
+
+public class LostPostQueryService {
+
+}
