+++ src/main/java/kr/co/pawong/pwbe/lostPost/domain/LostPostUpdate.java
@@ -1,0 +1,5 @@
+package kr.co.pawong.pwbe.lostPost.domain;
+
+public class LostPostUpdate {
+
+}
