+++ src/main/java/kr/co/pawong/pwbe/lostPost/enums/PostStatus.java
@@ -1,0 +1,6 @@
+package kr.co.pawong.pwbe.lostPost.enums;
+
+public enum PostStatus {
+    ACTIVE,
+    DELETED
+}
