+++ src/main/java/kr/co/pawong/pwbe/lostPost/enums/PostType.java
@@ -1,0 +1,7 @@
+package kr.co.pawong.pwbe.lostPost.enums;
+
+public enum PostType {
+    LOST,
+    FOUND,
+    FOSTER
+}
